requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
azure-storage-blob==12.24.1
fastapi==0.103.2
locust==2.32.6
numpy==2.2.3
pandas==2.2.2
Pillow==10.0.1
pydantic==2.7.4
Requests==2.32.2
rich==13.6.0
starlette==0.27.0
torch==2.3.1+cpu
torchvision==0.18.1+cpu
six==1.17.0
uvicorn==0.23.2
validators==0.20.0